Add out side temp

+++ app/src/main/java/com/cyberello/ksfarm/data/json/TemperatureData.java
@@ -4,5 +4,5 @@ public class TemperatureData {
 
     public double temperature;
     public double humidity;
-    public int pressure;
+    public double pressure;
 }
